requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
firebase-admin>=6.2.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
aiohttp>=3.8.5
python-multipart>=0.0.6
asyncio>=3.4.3
httpx>=0.25.0
pytest>=7.4.2
pytest-asyncio>=0.21.1